# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Management API",
    "description": "CSE 341 Library Management API Documentation",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}